requirements.txt
Django==3.1
django-cors-headers==3.4.0
django-enum-choices==2.1.2
django-rest-marshmallow==4.0.2
django-storages==1.9.1
djangorestframework==3.11.0
marshmallow==3.6.1
pytest==5.3.5
pytest-django==3.9.0
pytest-env
gevent
gunicorn
redis
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
marshmallow-enum
django-redis
requests
whitenoise
django-oauth-toolkit
Babel
